requirements.txt
google-cloud
Flask==0.12.2
Werkzeug<0.13.0,>=0.12.0
